+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Changelog for Flamenco Manager
   completed, etc. the worker cannot update it any more.
 - Log activity when task gets cancelled by request of Flamenco Server.
 - Tweaked the colour scheme of the web interface to be a bit more muted and easier to read.
+- Limit latest image system queue to 3 images, and discard newer ones until the queue shrinks.
 
 
 ## Version 2.0.15 (released 2017-09-09)